+++ src/oracledb/impl/thick/var.pyx
@@ -1,5 +1,5 @@
 #------------------------------------------------------------------------------
-# Copyright (c) 2020, 2022, Oracle and/or its affiliates.
+# Copyright (c) 2020, 2023, Oracle and/or its affiliates.
 #
 # This software is dual-licensed to you under the Universal Permissive License
 # (UPL) 1.0 as shown at https://oss.oracle.com/licenses/upl and Apache License
@@ -276,3 +276,5 @@ cdef class ThickVarImpl(BaseVarImpl):
             if self.outconverter is not None:
                 value = self.outconverter(value)
             return value
+        elif self.convert_nulls:
+            return self.outconverter(None)